+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* New option for automatic legend display at startup, Lizmap Web Client 3.6
 * Add OpenTopoMap background for Lizmap Web Client 3.6
 * Always export the new checkbox for a fixed scale in the overview map
 * Remove the code related to the server side. It's a new plugin called "Lizmap server" now.